rakuten_seo_tool/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
stripe==11.4.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==11.0.0
celery==5.4.0
redis==5.2.1
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-allauth==65.3.0